🦄 refactor: 提取播放模式相关逻辑到新的类
